fix: remove unneeded files
